fix: resolve all compiler warnings (zero-warning build)
- crackalack_gen.c: cast size_t/unsigned long args to uint64_t for
  PRIu64 format specifiers (gws, kernel sizes, strlen, sizeof)
- crackalack_lookup.c: same casts for gws, sizeof(gpu_ulong) expressions
- rtc_decompress.c: cast sizeof expression to uint64_t
- crackalack_unit_tests.c: move 'err' variable into the #else (OpenCL)
  block where it is actually used; eliminated unused-variable warning
  in Metal builds
- test_index_to_plaintext.c: cast strlen() results to uint64_t
- test_index_to_plaintext_ntlm9.c: same

+++ test_index_to_plaintext_ntlm9.c
@@ -65,7 +65,7 @@ int gpu_test_index_to_plaintext_ntlm9(gpu_device device, gpu_context context, gp
   if (strcmp(expected_plaintext, (char *)plaintext) == 0)
     test_passed = 1;
   else {
-    printf("\n\nGPU error:\n\tIndex: %"PRIu64"\n\tExpected: [%"PRIu64"][%s]\n\tCalculated: [%s]\n\n", index, strlen(expected_plaintext), expected_plaintext, plaintext);
+    printf("\n\nGPU error:\n\tIndex: %"PRIu64"\n\tExpected: [%"PRIu64"][%s]\n\tCalculated: [%s]\n\n", index, (uint64_t)strlen(expected_plaintext), expected_plaintext, plaintext);
   }
 
   CLFREEBUFFER(index_buffer);
